Moved classes to new namespace
